rename secondary to sequence
