+++ crates/astria-sequencer/src/lib.rs
@@ -5,7 +5,7 @@ pub(crate) mod component;
 pub mod config;
 pub(crate) mod crypto;
 pub(crate) mod genesis;
-pub(crate) mod secondary;
+pub(crate) mod sequence;
 mod sequencer;
 pub(crate) mod service;
 pub(crate) mod state_ext;
